+++ src/cmd/crud.ts
@@ -27,11 +27,16 @@ const run = async () => {
     config();
 
     const argv = await yargs(hideBin(process.argv))
-        .config({ schemaGlob: "./src/**/*.graphql", serverAddress: "127.0.0.1:9000" })
+        .config({
+            schemaGlob: "./src/**/*.graphql",
+            serverAddress: "127.0.0.1:9000",
+            namespace: "",
+        })
         .pkgConf("crud").argv;
 
     let schemaGlob: string = argv.schemaGlob as string;
     let serverAddress: string = argv.serverAddress as string;
+    let namespace: string = argv.namespace as string;
 
     if (process.env.CRUD_SCHEMA_GLOB) {
         schemaGlob = process.env.CRUD_SCHEMA_GLOB;
@@ -41,18 +46,29 @@ const run = async () => {
         serverAddress = process.env.CRUD_SERVER_ADDRESS;
     }
 
+    if (process.env.CRUD_NAMESPACE) {
+        namespace = process.env.CRUD_NAMESPACE;
+    }
+
+    if (namespace === "Fraym") {
+        throw new Error("Cannot use Fraym as namespace prefix as it is reserved for fraym apps");
+    }
+
     const schema = await loadSchema(`${schemaGlob}`, {
         loaders: [new GraphQLFileLoader()],
     });
 
-    const definitions = getTypeDefinition(schema);
+    const definitions = getTypeDefinition(schema, namespace);
 
-    await migrateSchemas(definitions, serverAddress);
+    await migrateSchemas(definitions, serverAddress, namespace);
 };
 
 run();
 
-const getTypeDefinition = (schema: GraphQLSchema): Record<string, TypeDefinition> => {
+const getTypeDefinition = (
+    schema: GraphQLSchema,
+    namespace: string
+): Record<string, TypeDefinition> => {
     const definitions: Record<string, TypeDefinition> = {};
 
     schema.toConfig().types.forEach(t => {
@@ -60,7 +76,8 @@ const getTypeDefinition = (schema: GraphQLSchema): Record<string, TypeDefinition
             return;
         }
 
-        const name = t.toString();
+        const name = `${namespace}${t.toString()}`;
+        ensureValidName(name);
 
         if (definitions[name]) {
             throw new Error(
@@ -69,12 +86,12 @@ const getTypeDefinition = (schema: GraphQLSchema): Record<string, TypeDefinition
         }
 
         if (t instanceof GraphQLObjectType) {
-            definitions[name] = getTypeDefinitionFromGraphQLObjectType(t);
+            definitions[name] = getTypeDefinitionFromGraphQLObjectType(t, namespace);
             return;
         }
 
         if (t instanceof GraphQLEnumType) {
-            definitions[name] = getTypeDefinitionFromGraphQLEnumType(t);
+            definitions[name] = getTypeDefinitionFromGraphQLEnumType(t, namespace);
             return;
         }
     });
@@ -82,8 +99,13 @@ const getTypeDefinition = (schema: GraphQLSchema): Record<string, TypeDefinition
     return definitions;
 };
 
-const getTypeDefinitionFromGraphQLEnumType = (t: GraphQLEnumType): TypeDefinition => {
-    const name = t.toString();
+const getTypeDefinitionFromGraphQLEnumType = (
+    t: GraphQLEnumType,
+    namespace: string
+): TypeDefinition => {
+    const name = `${namespace}${t.toString()}`;
+    ensureValidName(name);
+
     let enumValuesString = "";
 
     t.astNode?.values?.forEach(value => {
@@ -99,14 +121,17 @@ const getTypeDefinitionFromGraphQLEnumType = (t: GraphQLEnumType): TypeDefinitio
     };
 };
 
-const getTypeDefinitionFromGraphQLObjectType = (t: GraphQLObjectType): TypeDefinition => {
+const getTypeDefinitionFromGraphQLObjectType = (
+    t: GraphQLObjectType,
+    namespace: string
+): TypeDefinition => {
     const isCrudType =
         (t.astNode?.directives &&
             t.astNode?.directives.length > 0 &&
             t.astNode.directives[0].name.value === "crudType") ??
         false;
 
-    const name = t.toString();
+    const name = `${namespace}${t.toString()}`;
     let objectDirectivesString = "";
     let objectFieldsString = "";
     let nestedTypes: string[] = [];
@@ -116,7 +141,7 @@ const getTypeDefinitionFromGraphQLObjectType = (t: GraphQLObjectType): TypeDefin
     });
 
     t.astNode?.fields?.forEach(f => {
-        const { str, nestedTypes: newNestedTypes } = getFieldStringAndNestedTypes(f);
+        const { str, nestedTypes: newNestedTypes } = getFieldStringAndNestedTypes(f, namespace);
         objectFieldsString += str;
 
         newNestedTypes.forEach(nested => {
@@ -140,14 +165,14 @@ interface FieldData {
     nestedTypes: string[];
 }
 
-const getFieldStringAndNestedTypes = (f: FieldDefinitionNode): FieldData => {
+const getFieldStringAndNestedTypes = (f: FieldDefinitionNode, namespace: string): FieldData => {
     let directivesString = "";
 
     f.directives?.forEach(d => {
         directivesString += getDirectiveString(d);
     });
 
-    const { nestedType, str: typeString } = getTypeData(f.type);
+    const { nestedType, str: typeString } = getTypeData(f.type, namespace);
 
     const nestedTypes: string[] = [];
 
@@ -166,10 +191,11 @@ interface TypeData {
     nestedType?: string;
 }
 
-const getTypeData = (t: TypeNode): TypeData => {
+const getTypeData = (t: TypeNode, namespace: string): TypeData => {
     switch (t.kind) {
         case Kind.NAMED_TYPE:
             const name = t.name.value;
+            ensureValidName(`${namespace}${name}`);
 
             return name === "String" ||
                 name === "Float" ||
@@ -181,18 +207,21 @@ const getTypeData = (t: TypeNode): TypeData => {
                       str: name,
                   }
                 : {
-                      str: name,
-                      nestedType: name,
+                      str: `${namespace}${name}`,
+                      nestedType: `${namespace}${name}`,
                   };
         case Kind.LIST_TYPE:
-            const { nestedType: listNestedType, str: listStr } = getTypeData(t.type);
+            const { nestedType: listNestedType, str: listStr } = getTypeData(t.type, namespace);
 
             return {
                 str: `[${listStr}]`,
                 nestedType: listNestedType,
             };
         case Kind.NON_NULL_TYPE:
-            const { nestedType: nonNullNestedType, str: nonNullStr } = getTypeData(t.type);
+            const { nestedType: nonNullNestedType, str: nonNullStr } = getTypeData(
+                t.type,
+                namespace
+            );
 
             return {
                 str: `${nonNullStr}!`,
@@ -260,10 +289,13 @@ const getValueString = (v: ConstValueNode): string => {
 
 const migrateSchemas = async (
     definitions: Record<string, TypeDefinition>,
-    serverAddress: string
+    serverAddress: string,
+    namespace: string
 ) => {
     const managementClient = await newManagementClient({ serverAddress });
-    const existingTypeNames = await managementClient.getAllTypes();
+    const existingTypeNames = (await managementClient.getAllTypes()).filter(name =>
+        name.startsWith(namespace)
+    );
 
     let createSchema = "";
     let updateSchema = "";
@@ -333,5 +365,11 @@ const migrateSchemas = async (
         console.log(`Removing ${typesToRemove.length} types: ${typesToRemove}...`);
         await managementClient.removeTypes(typesToRemove).catch(console.log);
         console.log(`Removed ${typesToRemove.length} types`);
+    }
+};
+
+const ensureValidName = (name: string): void | never => {
+    if (name.startsWith("Fraym")) {
+        throw new Error("Cannot use Fraym as type name prefix as it is reserved for fraym apps");
     }
 };